requirements.txt
anyio==1.4.0
asgiref==3.2.10
async-generator==1.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-braces==1.14.0
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-model-utils==4.0.0
django-random-queryset==0.1.3
django-robots==4.0
document==1.0
h11==0.9.0
httpcore==0.10.1
httpx==0.14.1
idna==2.10
oauthlib==3.1.0
Pillow==7.2.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
sitemaps==0.1.0
six==1.15.0
sniffio==1.1.0
sqlparse==0.3.1
swapper==1.1.2.post1
urllib3==1.25.10
